requirements.txt
flask
gunicorn
python-dotenv
pyjwt
supabase
